+++ frontend/nuxt.config.ts
@@ -28,7 +28,8 @@ export default defineNuxtConfig({
 
 	runtimeConfig: {
 		public: {
-			apiBase: process.env.NUXT_PUBLIC_API_BASE || 'http://localhost:8000/api'
+			apiBase: process.env.NUXT_PUBLIC_API_BASE || 'http://localhost:8000/api',
+			telegramBotId: process.env.NUXT_PUBLIC_TELEGRAM_BOT_ID || ''
 		}
 	},